+++ spec/controllers/application_controller_spec.rb
@@ -36,7 +36,7 @@ RSpec.describe ApplicationController, type: :controller do
     describe "when a new version of the application is deployed" do
 
       it "busts the ETAG" do
-        controller.stub(:etag_seed).and_return "CURRENT_SHA1"
+        allow(controller).to receive(:etag_seed).and_return("CURRENT_SHA1")
 
         get :show, :id => "ignored"
         assert_response 200, @response.body
@@ -46,7 +46,7 @@ RSpec.describe ApplicationController, type: :controller do
         assert_response 304, @response.body
 
         # Deploy!
-        controller.stub(:etag_seed).and_return "NEW_SHA1"
+        allow(controller).to receive(:etag_seed).and_return "NEW_SHA1"
 
         @request.env["HTTP_IF_NONE_MATCH"] = etag
         get :show, :id => "ignored"
@@ -56,7 +56,7 @@ RSpec.describe ApplicationController, type: :controller do
 
     context "when a flash message is shown" do
       before(:each) do
-        controller.stub(:flash).and_return({"message" => "I am not empty"})
+        allow(controller).to receive(:flash).and_return({"message" => "I am not empty"})
       end
 
       include_examples "does not generate an ETAG"